move groupBy helper to common
